Create Book Store
Private members of the class BOOK are
BOOK_NO	integer type
BOOK_TITLE	20 characters
PRICE			float(price per copy)
TOTAL—COST( A function to calculate the total cost for N number of copies, where N is passed to the function as argument
Public members of the class BOOK are
INPUT()	Function to read BOOK_NO, BOOK_TITLE, PRICE
PURCHASE(} Function to ask the user to input the number of copies to be purchased. It invokes TOTAL COST( ) and prints the total cost to be paid by the user.